Move screens into their own respective "module"
